feat: Add controller to update 'disable' field based on user type
This commit adds a new controller that updates the 'disable' field in user models
such as Admin, Callcenter, Finance, SuperAdmin, DeliveryGuy, and Staff. The controller
uses the 'updateDisableField' function to perform the update, and the user type is
determined using a switch statement. The updated user is sent as a response upon
successful update.

# [JIRA-123] or [Issue Number] - if applicable